+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `bongolivesms-laravel` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.0.0 - 2020-11-27
 
-- initial release
+- initial release with ability to only send one message at a time 
+
+
+## 1.0.1 - 2020-12-05
+
+- Ability to send multiple msisdn on one call  send__multiple_recipient($source_addr,$message,$recipient_array)